+++ docs/changelog.md
@@ -13,6 +13,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 - Add the ability to externally manage pip with the `PYAPP_PIP_EXTERNAL` and `PYAPP_PIP_VERSION` options
 - Allow for project installation with a dependency file using the `PYAPP_PROJECT_DEPENDENCY_FILE` option
 - Add management command to directly invoke pip with the installed Python
+- Add management command to output the path to the installed Python
 
 ***Fixed:***